Added some functionality

+++ Controllers/ProductController.cs
@@ -1,6 +1,9 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Data;
+using System.Security.Claims;
 using WebShop.API.Models.Domain;
 using WebShop.API.Models.Dto;
 using WebShop.API.Services.Interfaces;
@@ -20,10 +23,13 @@ namespace WebShop.API.Controllers
             this.mapper = mapper;
         }
 
+        [Authorize(Roles = "Manager")]
         [HttpPost]
         public async Task<IActionResult> CreateProduct([FromBody] CreateProductRequestDto createProductRequestDto)
         {
             var productDomain = mapper.Map<Product>(createProductRequestDto);
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+            productDomain.CreatedBy = userId;
 
             productDomain = await productService.CreateProductAsync(productDomain);
 
@@ -32,14 +38,16 @@ namespace WebShop.API.Controllers
             return CreatedAtAction(nameof(GetProductById), new { productId = productDto.ProductId }, productDto);
         }
 
+        [Authorize(Roles = "Admin, Manager, RegularUser")]
         [HttpGet]
-        public async Task<IActionResult> GetAllProducts()
+        public async Task<IActionResult> GetAllProducts([FromQuery] string? filterOn, [FromQuery] string? filterQuery, [FromQuery] string? sortBy, [FromQuery] bool? isAscending, int pageNumber=1, int pageSize=1000)
         {
-            var productsDomain = await productService.GetAllProductsAsync();
+            var productsDomain = await productService.GetAllProductsAsync(filterOn, filterQuery, sortBy, isAscending ?? true, pageNumber, pageSize);
             var productsDto = mapper.Map<List<ProductDto>>(productsDomain);
             return Ok(productsDto);
         }
 
+        [Authorize(Roles = "Admin, Manager, RegularUser")]
         [HttpGet]
         [Route("{productId:Guid}")]
         public async Task<IActionResult> GetProductById([FromRoute] Guid productId)
@@ -47,7 +55,7 @@ namespace WebShop.API.Controllers
             var productDomain = await productService.GetProductByIdAsync(productId);
             if (productDomain == null)
             {
-                return NotFound();
+                return NotFound("Proizvod ne postoji");
             }
 
             var productDto = mapper.Map<ProductDto>(productDomain);
@@ -55,33 +63,48 @@ namespace WebShop.API.Controllers
             return Ok(productDto);
         }
 
+        [Authorize(Roles = "Admin, Manager")]
         [HttpDelete]
         [Route("{productId:Guid}")]
         public async Task<IActionResult> DeleteProduct([FromRoute] Guid productId)
         {
-            var productDomain = await productService.DeleteProductAsync(productId);
-            if (productDomain == null)
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+            var roles = User.FindAll(ClaimTypes.Role).Select(r => r.Value).ToList();
+
+            var product = await productService.GetProductByIdAsync(productId);
+            if (product == null) return NotFound();
+
+            if (roles.Contains("Admin") || (roles.Contains("Manager") && product.CreatedBy == userId))
             {
-                return NotFound();
+                var deletedProduct = await productService.DeleteProductAsync(productId);
+                var productDto = mapper.Map<ProductDto>(deletedProduct);
+                return Ok(productDto);
             }
-            var productDto = mapper.Map<ProductDto>(productDomain);
-            return Ok(productDto);
+            return Forbid("Nemate dozvolu da obrišete ovaj proizvod.");
         }
 
+        [Authorize(Roles = "Admin, Manager")]
         [HttpPut]
         [Route("{productId:Guid}")]
         public async Task<IActionResult> UpdateProduct([FromRoute] Guid productId, [FromBody] UpdateProductRequestDto updateProductRequestDto)
         {
-            var productDomain = mapper.Map<Product>(updateProductRequestDto);
-            productDomain = await productService.UpdateProductAsync(productId, productDomain);
 
-            if (productDomain == null)
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+            var roles = User.FindAll(ClaimTypes.Role).Select(r => r.Value).ToList();
+
+            var product = await productService.GetProductByIdAsync(productId);
+            if (product == null) return NotFound();
+
+            if (roles.Contains("Admin") || ((roles.Contains("Manager") && product.CreatedBy == userId)))
             {
-                return NotFound();
+                var updatedProductDomain = mapper.Map<Product>(updateProductRequestDto);
+                updatedProductDomain.CreatedBy = product.CreatedBy;
+
+                var updatedProduct = await productService.UpdateProductAsync(productId, updatedProductDomain);
+                var productDto = mapper.Map<ProductDto>(updatedProduct);
+                return Ok(productDto);
             }
-            //map to dto
-            var productDto = mapper.Map<ProductDto>(productDomain);
-            return Ok(productDto);
+            return Forbid("Nemate dozvolu da obrišete ovaj proizvod.");
         }
     }
 }
